% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.matrix.R
\name{combine_matrix}
\alias{combine_matrix}
\title{Funciones genericas}
\usage{
combine_matrix(matrix.odd, matrix.even)
}
\arguments{
\item{matrix.odd}{matrix.odd}

\item{matrix.even}{matrix.even}
}
\description{
Combine Matrix for design matrix of random effects
}
