% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained.mixed.model.mcmc.R
\name{constrained.mixed.model.mcmc}
\alias{constrained.mixed.model.mcmc}
\title{Muestreo de Gibbs para el modelo mixto restringido}
\usage{
constrained.mixed.model.mcmc(n.iter, data)
}
\arguments{
\item{n.iter}{number of iterations}

\item{data}{data object with observations}
}
\description{
mcmc
}
