% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.generation.R
\name{data.generation}
\alias{data.generation}
\title{Generacion de datos simulados}
\usage{
data.generation(n.fixed.effects, n.random.effects, n.sample)
}
\arguments{
\item{n.fixed.effects}{number of fixed effects}

\item{n.random.effects}{n random effects}

\item{n.sample}{size of sample}
}
\description{
generate data
}
