% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{data_generation}
\alias{data_generation}
\title{Generacion de datos simulados}
\usage{
data_generation(p, m, n)
}
\arguments{
\item{p}{number of fix factors}

\item{m}{number of random factors}

\item{n}{size of sample}
}
\description{
generate data
}
\examples{
data <- data_generation(p=10, m=10, n=100)
}
